Move passphrase related code to a folder
