+++ src/components/passphrase/passphrase.less
