+++ src/liskNano.js
@@ -6,7 +6,8 @@ import './components/main/main';
 import './components/main/setSecondPassService';
 import './components/main/setSecondPassDirective';
 import './components/login/login';
-import './components/login/passphrase';
+import './components/passphrase/passphrase';
+import './components/passphrase/passphraseService';
 import './components/top/top';
 import './components/header/header';
 import './components/send/send';
@@ -25,7 +26,6 @@ import './components/delegateRegistration/delegateRegistration';
 import './services/api/peers';
 import './services/lsk';
 import './services/dialog';
-import './services/passphrase';
 import './services/signVerify';
 import './services/account';
 import './services/api/delegateApi';
